+++ src/Options/OptionContainer.php
@@ -3,7 +3,7 @@
 namespace MAChitgarha\Json\Options;
 
 /**
- * Every class containing options for specific functionality must extend this class.
+ * Every class containing options for specific operation must extend this class.
  */
 abstract class OptionContainer
 {